cabravietnamirachamsinpeladobackend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.2
psycopg2-binary==2.9.7
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
Pillow==10.4.0
gunicorn==21.2.0
setuptools>=65.5.1
